Fix stop error, fix dissipate, cleanup
The animate module was trying to call heatmap:render after it had
already shut down the heatmap process. Put in a check to make sure the
heatmap is still alive before trying to render it.

Dissipate was just plain weird. Reset the dissipate on each remove to
dissipate the remaining heat.

Took out some of the io:format noise. It's useful when first getting all
the plumbing done but annoying later on.