WIP: Update algo values from webpage
I want to play with the algorithm inputs from the web page,
interactively, so that I don't have to stop the app and recompile
everytime I want to test out a different value.

Currently only the "Falloff" (How fast the heat added is reduced for
each concentric square away from the center) and "Range" (how many
concentric squares of heat to track) are editable.